move github settings type to it's own file
