+++ src/issue-filing/services/github/github-issue-filing-service.tsx
@@ -9,12 +9,9 @@ import { createSettingsGetter } from '../../common/create-settings-getter';
 import { IssueFilingService } from '../../types/issue-filing-service';
 import { SettingsFormProps } from '../../types/settings-form-props';
 import { gitHubIssueFilingUrlProvider } from './create-github-issue-filing-url';
+import { GitHubIssueFilingSettings } from './github-issue-filing-settings';
 
 const GitHubIssueFilingServiceKey = 'gitHub';
-
-export type GitHubIssueFilingSettings = {
-    repository: string;
-};
 
 function buildStoreData(repository: string): GitHubIssueFilingSettings {
     return {
