+++ src/issue-filing/services/github/github-issue-filing-settings.tsx
@@ -1,0 +1,5 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+export type GitHubIssueFilingSettings = {
+    repository: string;
+};
