+++ src/tests/unit/tests/issue-filing/services/github/github-issue-filing-service.test.tsx
@@ -9,10 +9,8 @@ import { IssueFilingServicePropertiesMap } from '../../../../../../common/types/
 import { SettingsDeps } from '../../../../../../DetailsView/components/settings-panel/settings/settings-props';
 import { OnPropertyUpdateCallback } from '../../../../../../issue-filing/components/issue-filing-settings-container';
 import { gitHubIssueFilingUrlProvider } from '../../../../../../issue-filing/services/github/create-github-issue-filing-url';
-import {
-    GitHubIssueFilingService,
-    GitHubIssueFilingSettings,
-} from '../../../../../../issue-filing/services/github/github-issue-filing-service';
+import { GitHubIssueFilingService } from '../../../../../../issue-filing/services/github/github-issue-filing-service';
+import { GitHubIssueFilingSettings } from '../../../../../../issue-filing/services/github/github-issue-filing-settings';
 import { SettingsFormProps } from '../../../../../../issue-filing/types/settings-form-props';
 
 describe('GithubIssueFilingServiceTest', () => {
